+++ Source/SwiftSky.swift
@@ -267,12 +267,11 @@ public struct SwiftSky {
             }
         }
         
-        // get api url
-        let url = ApiUrl(located, date: date, exclude: exclude).url
+        guard let url = ApiUrl(located, date: date, exclude: exclude).url else {
+            return completion(.failure(.noApiKey))
+        }
         
-        // perform request
-        guard url != nil else { return completion(.failure(.noApiKey)) }
-        request(url!, method: .get, headers: ["Accept-Encoding":"gzip"]).responseJSON { response in
+        request(url, method: .get, headers: ["Accept-Encoding":"gzip"]).responseJSON { response in
             switch response.result {
             case .success(let data):
                 completion(.success(Forecast(data, headers: response.response?.allHeaderFields)))